Move Printer class inside Printer module
