+++ src/php/Exceptions/HtmlExceptionPrinter.php
@@ -7,7 +7,7 @@ namespace Phel\Exceptions;
 use Phel\Compiler\Emitter\OutputEmitter\Munge;
 use Phel\Compiler\ReadModel\CodeSnippet;
 use Phel\Lang\IFn;
-use Phel\Printer;
+use Phel\Printer\Printer;
 use ReflectionClass;
 use Throwable;
