+++ src/php/Printer/ElementPrinter/PhelArrayPrinter.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Phel\Printer\ElementPrinter;
 
 use Phel\Lang\PhelArray;
-use Phel\Printer;
+use Phel\Printer\Printer;
 
 /**
  * @implements ElementPrinterInterface<PhelArray>
